+++ rwengine/src/render/TextRenderer.cpp
@@ -197,7 +197,7 @@ void TextRenderer::renderText(const TextRenderer::TextInfo& ti)
 
 	auto text = ti.text;
 	
-	for( int i = 0; i < text.length(); ++i )
+	for (size_t i = 0; i < text.length(); ++i)
 	{
 		char c = text[i];
